requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
opencv-python==4.9.0.80
matplotlib==3.8.3
seaborn==0.13.2
pandas==2.2.1
scikit-learn==1.4.1.post1
imageio==2.34.0
google-generativeai==0.3.2
fpdf2==2.7.8 